# Docs/MultBot-OpenAPI-3.0.yaml
openapi: 3.0.3
info:
  title: MultBot API
  description: |
    API para gerenciamento de múltiplos bots Telegram de pagamento com integração DEPIX.
    
    ## Autenticação
    A API utiliza JWT Bearer Token. Após login, inclua o token em todas as requisições:
    ```
    Authorization: Bearer <token>
    ```
  version: 1.0.0
  contact:
    name: MultBot Team

servers:
  - url: http://localhost:3000/api
    description: Servidor de desenvolvimento
  - url: https://api.multbot.com/api
    description: Servidor de produção

tags:
  - name: Auth
    description: Autenticação e autorização
  - name: Dashboard
    description: KPIs e estatísticas agregadas
  - name: Bots
    description: Gerenciamento de bots Telegram
  - name: Transactions
    description: Histórico e gerenciamento de transações
  - name: Settings
    description: Configurações globais do sistema

paths:
  # ============================================
  # AUTH
  # ============================================
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login do administrador
      description: Autentica o administrador e retorna JWT token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: admin@test.com
              password: password123
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Email ou senha inválidos

  # ============================================
  # DASHBOARD
  # ============================================
  /dashboard/stats:
    get:
      tags:
        - Dashboard
      summary: Estatísticas do dashboard
      description: Retorna KPIs agregados de todos os bots e transações
      operationId: getDashboardStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Estatísticas retornadas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ============================================
  # BOTS
  # ============================================
  /bots:
    get:
      tags:
        - Bots
      summary: Listar todos os bots
      description: Retorna lista de todos os bots cadastrados com suas transações
      operationId: listBots
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filtrar por status
          schema:
            type: string
            enum: [active, inactive]
        - name: search
          in: query
          description: Buscar por nome, owner ou telegram
          schema:
            type: string
      responses:
        '200':
          description: Lista de bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bot'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Bots
      summary: Criar novo bot
      description: Cria um novo bot Telegram para um logista
      operationId: createBot
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBotRequest'
      responses:
        '201':
          description: Bot criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: Token do Telegram já cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Token do Telegram já está em uso

  /bots/{id}:
    get:
      tags:
        - Bots
      summary: Detalhes de um bot
      description: Retorna informações detalhadas de um bot específico
      operationId: getBot
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BotId'
      responses:
        '200':
          description: Detalhes do bot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotDetails'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags:
        - Bots
      summary: Atualizar bot
      description: Atualiza informações de um bot existente
      operationId: updateBot
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBotRequest'
      responses:
        '200':
          description: Bot atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Bots
      summary: Remover bot
      description: Remove um bot do sistema (soft delete - marca como inativo)
      operationId: deleteBot
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BotId'
      responses:
        '204':
          description: Bot removido com sucesso
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ============================================
  # TRANSACTIONS
  # ============================================
  /transactions:
    get:
      tags:
        - Transactions
      summary: Listar transações
      description: Retorna lista paginada de transações com filtros
      operationId: listTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Número da página (1-indexed)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Itens por página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Filtrar por status
          schema:
            type: string
            enum: [processing, completed, failed]
        - name: botId
          in: query
          description: Filtrar por bot
          schema:
            type: string
        - name: search
          in: query
          description: Buscar por ID, bot ou cliente
          schema:
            type: string
        - name: dateFrom
          in: query
          description: Data inicial (ISO 8601)
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          description: Data final (ISO 8601)
          schema:
            type: string
            format: date
        - name: sortBy
          in: query
          description: Campo para ordenação
          schema:
            type: string
            enum: [createdAt, amount, status]
            default: createdAt
        - name: sortOrder
          in: query
          description: Direção da ordenação
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Lista paginada de transações
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /transactions/{id}:
    get:
      tags:
        - Transactions
      summary: Detalhes de uma transação
      description: Retorna informações detalhadas de uma transação
      operationId: getTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Detalhes da transação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDetails'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /transactions/export:
    get:
      tags:
        - Transactions
      summary: Exportar transações
      description: Exporta transações em formato CSV
      operationId: exportTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [processing, completed, failed]
        - name: botId
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Arquivo CSV
          content:
            text/csv:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              example: attachment; filename="transactions_2026-02-05.csv"
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ============================================
  # SETTINGS
  # ============================================
  /settings:
    get:
      tags:
        - Settings
      summary: Obter configurações
      description: Retorna todas as configurações globais do sistema
      operationId: getSettings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dados de configuração
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsData'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    put:
      tags:
        - Settings
      summary: Atualizar configurações
      description: Atualiza parâmetros globais e credenciais API
      operationId: updateSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsData'
      responses:
        '200':
          description: Configurações atualizadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsData'
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /settings/test-depix:
    post:
      tags:
        - Settings
      summary: Testar conexão Depix
      description: Verifica se as credenciais da API Depix são válidas
      operationId: testDepix
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Resultado do teste
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'

components:
  # ============================================
  # SECURITY SCHEMES
  # ============================================
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido via /auth/login

  # ============================================
  # PARAMETERS
  # ============================================
  parameters:
    BotId:
      name: id
      in: path
      required: true
      description: ID único do bot
      schema:
        type: string
        example: clxyz123abc

    TransactionId:
      name: id
      in: path
      required: true
      description: ID único da transação
      schema:
        type: string
        example: TX-2026-0342

  # ============================================
  # RESPONSES
  # ============================================
  responses:
    UnauthorizedError:
      description: Token ausente ou inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Token obrigatório

    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Recurso não encontrado

  # ============================================
  # SCHEMAS
  # ============================================
  schemas:
    # --- Auth ---
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Email do administrador
        password:
          type: string
          format: password
          minLength: 6
          description: Senha do administrador

    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT token para autenticação
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        admin:
          type: object
          properties:
            email:
              type: string
              example: admin@test.com
            name:
              type: string
              example: Administrador

    # --- Dashboard ---
    DashboardStats:
      type: object
      properties:
        botsCount:
          type: integer
          description: Total de bots cadastrados
          example: 5
        transactionsCount:
          type: integer
          description: Total de transações
          example: 1234
        totalRevenue:
          type: integer
          description: Receita total em centavos
          example: 15000000
        successRate:
          type: number
          format: float
          description: Taxa de sucesso (0-100)
          example: 94.5
        topBots:
          type: array
          description: Top 5 bots por receita
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              revenue:
                type: integer
                description: Receita em centavos

    # --- Bot ---
    Bot:
      type: object
      properties:
        id:
          type: string
          example: clxyz123abc
        name:
          type: string
          example: Bot TechStore
        telegramToken:
          type: string
          description: Token do bot (parcialmente oculto)
          example: "123456:ABC***XYZ"
        telegramUsername:
          type: string
          example: "@techstore_depix_bot"
        ownerName:
          type: string
          description: Nome do logista
          example: João Silva
        depixAddress:
          type: string
          description: Endereço Liquid para recebimento
          example: VJL...abc
        splitRate:
          type: number
          format: float
          description: Taxa da plataforma (0-1)
          example: 0.10
        status:
          type: string
          enum: [active, inactive]
          example: active
        totalRevenue:
          type: integer
          description: Receita total em centavos
          example: 4528000
        transactionsCount:
          type: integer
          description: Número de transações
          example: 342
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    BotDetails:
      allOf:
        - $ref: '#/components/schemas/Bot'
        - type: object
          properties:
            transactions:
              type: array
              description: Últimas 10 transações
              items:
                $ref: '#/components/schemas/TransactionSummary'
            stats:
              type: object
              properties:
                todayRevenue:
                  type: integer
                weekRevenue:
                  type: integer
                monthRevenue:
                  type: integer
                avgTicket:
                  type: integer

    CreateBotRequest:
      type: object
      required:
        - name
        - telegramToken
        - ownerName
        - depixAddress
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          description: Nome do bot
          example: Bot TechStore
        telegramToken:
          type: string
          description: Token do bot do BotFather
          example: "123456789:ABCdefGHIjklMNOpqrsTUVwxyz"
        ownerName:
          type: string
          minLength: 3
          maxLength: 100
          description: Nome do logista/proprietário
          example: João Silva
        depixAddress:
          type: string
          description: Endereço Liquid do logista
          example: VJLxxxxxxxxxxxxxxxxxxxxx
        splitRate:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.10
          description: Taxa da plataforma (padrão 10%)

    UpdateBotRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        ownerName:
          type: string
          minLength: 3
          maxLength: 100
        depixAddress:
          type: string
        splitRate:
          type: number
          format: float
          minimum: 0
          maximum: 1
        status:
          type: string
          enum: [active, inactive]

    # --- Transaction ---
    Transaction:
      type: object
      properties:
        id:
          type: string
          example: TX-2026-0342
        botId:
          type: string
        botName:
          type: string
          example: Bot TechStore
        customerName:
          type: string
          example: Cliente Final A
        amountBrl:
          type: integer
          description: Valor em centavos
          example: 125000
        depixAmount:
          type: integer
          description: Valor Depix em satoshis
          example: 37500
        merchantSplit:
          type: integer
          description: Valor do logista em centavos
          example: 112500
        adminSplit:
          type: integer
          description: Valor da plataforma em centavos
          example: 12500
        pixKey:
          type: string
          description: Chave PIX utilizada (truncada)
          example: "abc123...xyz"
        status:
          type: string
          enum: [processing, completed, failed]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true

    TransactionSummary:
      type: object
      properties:
        id:
          type: string
        amountBrl:
          type: integer
        status:
          type: string
          enum: [processing, completed, failed]
        createdAt:
          type: string
          format: date-time

    TransactionDetails:
      allOf:
        - $ref: '#/components/schemas/Transaction'
        - type: object
          properties:
            bot:
              $ref: '#/components/schemas/Bot'
            depixTxId:
              type: string
              description: ID da transação na rede Liquid
            pixTxId:
              type: string
              description: ID da transação PIX

    TransactionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        pagination:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            total:
              type: integer
              description: Total de itens
              example: 1234
            totalPages:
              type: integer
              example: 62

    # --- Common ---
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Mensagem de erro
        details:
          type: object
          description: Detalhes adicionais (validação)
          additionalProperties: true

    # --- Settings ---
    SettingsData:
      type: object
      properties:
        defaultSplitRate:
          type: number
          format: float
          description: Taxa de split padrão (0-1)
          example: 0.10
        minSplitRate:
          type: number
          format: float
          description: Taxa mínima permitida
          example: 0.05
        maxSplitRate:
          type: number
          format: float
          description: Taxa máxima permitida
          example: 0.30
        depixApiUrl:
          type: string
          description: URL base da API Depix
          example: https://api.depix.com/v1
        depixApiKey:
          type: string
          description: Chave de API (não retornada por segurança)
        depixWebhookSecret:
          type: string
          description: Segredo do Webhook (não retornado por segurança)
        telegramApiId:
          type: string
          description: App ID do Telegram MTProto
        telegramApiHash:
          type: string
          description: App Hash do Telegram MTProto
        telegramPhone:
          type: string
          description: Telefone do bot admin
        notifyEmail:
          type: boolean
          description: Enviar email em falhas críticas
        notifyTelegram:
          type: boolean
          description: Enviar mensagem Telegram em falhas
        notifyMinAmount:
          type: integer
          description: Valor mínimo em centavos para notificação priorizada
          example: 100000